refactor: send deletion key via auth header, update deps

+++ frontend/src/routes/api/pastes/[id]/+server.ts
@@ -3,19 +3,23 @@ import { kitFetchWrapper } from '$lib/server/fetch';
 import { json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 
-export interface DeletePasteRequest {
-    deletionKey: string;
-}
-
 export const DELETE: RequestHandler = async ({ params, request, fetch }) => {
     const { id } = params;
-    const requestJson: DeletePasteRequest = await request.json();
+    const deletionKey = request.headers.get('authorization');
+    if (!deletionKey) {
+        return json({
+            success: false,
+            message: 'Missing deletion key'
+        }, {
+            status: 400
+        });
+    }
     await kitFetchWrapper(fetch, apiUrl(`pastes/${id}`), {
         method: 'DELETE',
         headers: {
-            'Content-Type': 'application/json'
+            'Content-Type': 'application/json',
+            "Authorization": deletionKey
         },
-        body: JSON.stringify(requestJson)
     });
     return json({ success: true });
 };
